feat(mesh-render): a malha desenha num SUB-RECTANGULO do alvo (`ScreenRect`)
O substrato dos quatro viewports da escultura (ordem do Enio,
2026-09-08). Ate' aqui os tre^s passes desta crate escreviam o alvo
INTEIRO, e quatro vistas ao mesmo tempo pedem quatro sub-rectangulos do
mesmo `game_rt`.

APPEND-ONLY: `render`, `render_gbuffer` e `render_ssao` ficam com a
assinatura que sempre tiveram e passam a DELEGAR com
`ScreenRect::full(size)`; os tre^s `*_in` novos recebem o `area`. Zero
churn nos chamadores, e o gate nº1 mede exactamente isso.

TRES METADES, e cada uma tem o seu modo de falha proprio:

  1. `set_viewport` E `set_scissor_rect`, os dois, por uma PORTA
     (`set_area`). O viewport TRANSFORMA e nao corta -- com ele sozinho a
     geometria que sai pelo lado do frustum continua a escrever no
     quadrante vizinho.
  2. O ASPECTO e o uniform de viewport sao os da VISTA, nunca os do alvo.
     E' a metade que uma implementacao apressada esquece, e ela e'
     invisivel numa vista so': com quatro quadrantes o alvo continua
     deitado e cada vista fica quase quadrada, entao o aspecto do alvo
     esticaria a peca nas QUATRO por igual -- que e' como um erro de
     escala deixa de se parecer com um erro.
  3. O uniform do SSAO ganha a ORIGEM, e o WGSL subtrai-a. O
     `@builtin(position)` do fragmento e' absoluto no framebuffer MESMO
     com `set_viewport`: sem isto um pixel do quadrante de baixo a`
     direita reconstruiria o NDC do canto do ecra e a oclusao sairia
     medida noutro sitio da peca. *Um `set_viewport` move o
     rasterizador e nao move a aritmetica do shader.*

A cor e' `LoadOp::Load` e a profundidade `Clear`, e a assimetria e'
load-bearing: a vista k+1 apaga a profundidade da k (um clear nao obedece
ao scissor) e pode -- a cor da k ja' foi escrita e ninguem volta a testar
profundidade contra ela.

QUATRO GATES DE GPU novos (`tests/gpu_viewport.rs`), e o primeiro e' o que
importa: `render_in` no alvo inteiro e' BYTE A BYTE o `render` de sempre
(0 de 131 072 bytes). Os outros tre^s: o scissor corta na coluna certa
(com a camera aproximada de proposito ate' a peca transbordar, senao o
gate mede uma esfera que ja' cabia), o aspecto e' o da vista (esfera a
`1,0123` numa vista quadrada de um alvo 2:1 -- com o defeito daria
`~0,5`), e duas vistas no mesmo quadro sobrevivem uma a` outra.

DOIS DEFEITOS DE REGUA apanhados de caminho, os dois pre'-existentes:

  * `SsaoRaw::SIZE` era gateado contra o LITERAL `96` com a frase
     «mat4x4 + 2 vec4». O campo novo reprovou-o, e o que ele pedia era que
     se actualizasse a frase -- nao que se conferisse nada. Passa a SOMAR
     os campos declarados no WGSL e a compara-los com o `size_of` do lado
     Rust, contando so' linhas de campo (um `///` que cite `vec4<f32>` na
     prosa nao e' um campo).
  * `coverage()` do `gpu_render` somava tre^s `u8`: em DEBUG entrava em
     panico (dai o `--release` no cabecalho do ficheiro, que era remedio
     do sintoma) e em RELEASE ENVOLVIA em silencio -- um pixel
     `(86, 85, 85)` soma 256, que modulo 256 e' 0, e a regua contava-o
     como FUNDO. Uma regua errada so' numa faixa estreita de valores e'
     pior que uma que falha sempre: ela passa em todo gate cuja peca nao
     calhe nessa faixa. Suite verde nos DOIS perfis depois da cura.

Porta nova na `Camera3d`: `view_axis` (o eixo alvo->olho, que o
`screen_basis` ja' derivava por dentro).

+++ crates/ph2d-mesh-render/tests/gpu_render.rs
@@ -272,12 +272,23 @@ fn lum(px: &[u8], x: u32, y: u32) -> f32 {
 }
 
 /// Fração de pixels que não são o fundo.
+///
+/// ⛔⛔ **A soma é em `u16`, e a versão em `u8` era um DEFEITO DE RÉGUA com dois
+/// sintomas** (achado 2026-09-08, ao correr esta suíte em debug): em **debug**
+/// ela entrava em pânico com *«attempt to add with overflow»* — daí o
+/// `--release` no cabeçalho deste ficheiro, que era um remédio do sintoma —, e
+/// em **release** ela ENVOLVIA em silêncio: um pixel `(86, 85, 85)` soma `256`,
+/// que módulo 256 é `0`, e a régua contava-o como **fundo**.
+///
+/// ⚠️ *Uma régua que dá a resposta errada só numa faixa estreita de valores é
+/// pior que uma que falha sempre: ela passa em todo gate cuja peça não calhe
+/// nessa faixa, e é o gate seguinte que herda o erro.*
 fn coverage(px: &[u8]) -> f32 {
     let lit = px
         .as_chunks::<4>()
         .0
         .iter()
-        .filter(|p| p[0] + p[1] + p[2] > 8)
+        .filter(|p| u16::from(p[0]) + u16::from(p[1]) + u16::from(p[2]) > 8)
         .count();
     lit as f32 / (W * H) as f32
 }
